+++ frontend/src/components/Main/Buttons/JoinCompetitionButton.tsx
