+++ frontend/src/pages/MainPage.tsx
@@ -1,13 +1,13 @@
 import { css } from '@style/css';
 
-import CreateCompetitionButton from '@/components/Commoms/Buttons/CreateCompetitionButton';
+import CreateCompetitionButton from '@/components/Main/Buttons/CreateCompetitionButton';
 import MainList from '@/components/Main/MainList';
 import { SITE } from '@/constants';
 
 function MainPage() {
   return (
     <main className={style}>
-      <span className={ProjectName}>{SITE.NAME} </span>
+      <span className={ProjectNameStyle}>{SITE.NAME} </span>
       <span>{SITE.PAGE_DESCRIPTION} </span>
       <CreateCompetitionButton />
       <MainList />
@@ -17,7 +17,7 @@ function MainPage() {
 
 export default MainPage;
 
-const ProjectName = css({
+const ProjectNameStyle = css({
   fontSize: '70px',
 });
